Some more formatting. Tried to slightly cleanup config loading.
Removed CustomBlockDefinitions section from config:
The ore names in the normal ProtectedBlocks section(s) can now directly
be id's as well. Also the data value (subid) can now be specified
together with the orename/id by separating it with a '~'. And the
default data value is now -1 (means it is ignored) instead of previously
0.