Update: Dropped support for arbitrary block ids.
It still tries to match block ids to existing known materials, but it no
longer works for 'unknown' materials. Previously this existed with mod
support in mind, but it is unknown if this was actually used / required
or even worked.

The preferred way to specify blocks is via material names only,
especially with the future MC 1.13 changes in mind (which will probably
require further changes to this, including complete removal of block id
matching, and sub-ids / data values).